requirements.txt
Django==2.2.4
django-braces==1.13.0
django-rest-framework==0.1.0
djangorestframework==3.10.2
pytz==2019.2
sqlparse==0.3.0
Unipath==1.1
